# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Dental Clinic Backend API
  version: 1.0.0
  description: >-
    Implemented backend contract. Dates and clock times are interpreted in the
    configured clinic timezone. Production cookie-authenticated operations also
    require an exact allowed Origin header.
servers:
  - url: /api/v1
security: []
...
paths:
...
  /appointments:
    get:
      tags: [Appointments]
      operationId: listAppointments
      description: Admin or receptionist.
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
        - { name: date, in: query, schema: { $ref: '#/components/schemas/LocalDate' } }
        - { $ref: '#/components/parameters/FromDate' }
        - { $ref: '#/components/parameters/ToDate' }
        - { name: dentistId, in: query, schema: { $ref: '#/components/schemas/ObjectId' } }
        - { name: serviceId, in: query, schema: { $ref: '#/components/schemas/ObjectId' } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/AppointmentStatus' } }
        - { name: phone, in: query, schema: { type: string, maxLength: 30 } }
      responses:
        '200': { $ref: '#/components/responses/AppointmentListSuccess' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
    post:
      tags: [Appointments]
      operationId: createPublicAppointment
      description: >-
        Public booking. Idempotency-Key is mandatory. Production also requires
        challengeToken. A successful retry with the same key and semantic body
        returns the original immutable public result with status 201; reuse with
        a different body returns 409. Successful records expire after the
        configured 1-168 hour idempotency window (24 hours by default), after
        which the key may create a new booking. Notification jobs are committed
        transactionally and delivered asynchronously; SMTP never controls booking
        acceptance.
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PublicBookingRequest' }
      responses:
        '201': { $ref: '#/components/responses/PublicBookingSuccess' }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /appointments/admin:
    post:
      tags: [Appointments]
      operationId: createAdminAppointment
      description: Admin or receptionist.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminBookingRequest' }
      responses:
        '201': { $ref: '#/components/responses/AppointmentSuccess' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /appointments/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Appointments]
      operationId: getAppointment
      description: Admin or receptionist.
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/AppointmentSuccess' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
  /appointments/{id}/availability:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Appointments, Scheduling]
      operationId: getRescheduleAvailability
      description: >-
        Admin or receptionist. Returns fresh availability while excluding only
        this appointment's current locks. The viewed appointment mutation
        version is required so stale reschedule editors fail closed.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: expectedMutationVersion, in: query, required: true, schema: { $ref: '#/components/schemas/AppointmentMutationVersion' } }
        - { name: dentistId, in: query, required: true, schema: { $ref: '#/components/schemas/ObjectId' } }
        - { name: serviceId, in: query, required: true, schema: { $ref: '#/components/schemas/ObjectId' } }
        - { name: date, in: query, required: true, schema: { $ref: '#/components/schemas/LocalDate' } }
      responses:
        '200': { $ref: '#/components/responses/AvailabilitySuccess' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /appointments/{id}/reschedule:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    patch:
      tags: [Appointments]
      operationId: rescheduleAppointment
      description: >-
        Admin or receptionist. Only pending or confirmed appointments may be
        rescheduled. The old lock remains on any failure; selecting the exact
        current occurrence is a concurrency-fenced no-op.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RescheduleRequest' }
      responses:
        '200': { $ref: '#/components/responses/AppointmentSuccess' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /appointments/{id}/status:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    patch:
      tags: [Appointments]
      operationId: updateAppointmentStatus
      description: Admin or receptionist; validated state transitions only.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AppointmentStatusRequest' }
      responses:
        '200': { $ref: '#/components/responses/AppointmentSuccess' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
  /appointments/{id}/cancel:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Appointments]
      operationId: cancelAppointment
      description: Admin or receptionist; releases booking and quota locks.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CancelAppointmentRequest' }
      responses:
        '200': { $ref: '#/components/responses/AppointmentSuccess' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ObjectId' }
...
    FromDate:
      name: from
      in: query
      schema: { $ref: '#/components/schemas/LocalDate' }
    ToDate:
      name: to
      in: query
      schema: { $ref: '#/components/schemas/LocalDate' }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Fresh UUIDv4 identifying one intended public booking. Reuse for the same
        semantic request replays the original successful public response within
        the configured idempotency window; reuse for a different request is 409.
      schema:
        type: string
        format: uuid
        pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
...
  responses:
...
    AvailabilitySuccess:
      description: Computed public availability
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [success, data]
            properties:
              success: { type: boolean, const: true }
              data:
                type: object
                additionalProperties: false
                required: [availability]
                properties:
                  availability: { $ref: '#/components/schemas/Availability' }
    PublicBookingSuccess:
      description: New or idempotently replayed public booking; both use status 201.
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [success, message, data]
            properties:
              success: { type: boolean, const: true }
              message: { type: string }
              data:
                type: object
                additionalProperties: false
                required: [appointment]
                properties:
                  appointment: { $ref: '#/components/schemas/AppointmentPublicResult' }
    AppointmentSuccess:
      description: Authorized appointment result containing administrative fields.
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [success, data]
            properties:
              success: { type: boolean, const: true }
              message: { type: string }
              data:
                type: object
                additionalProperties: false
                required: [appointment]
                properties:
                  appointment: { $ref: '#/components/schemas/AppointmentAdmin' }
    AppointmentListSuccess:
      description: Authorized paginated appointment list.
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [success, data]
            properties:
              success: { type: boolean, const: true }
              data:
                type: object
                additionalProperties: false
                required: [appointments, pagination]
                properties:
                  appointments:
                    type: array
                    items: { $ref: '#/components/schemas/AppointmentAdmin' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
...
    Error:
      description: Error response
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    ObjectId:
      type: string
      pattern: '^[0-9a-fA-F]{24}$'
    Slug:
      type: string
      minLength: 2
      maxLength: 180
      pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
...
    LocalDate:
      type: string
      pattern: '^\d{4}-\d{2}-\d{2}$'
      examples: ['2026-08-20']
    LocalTime:
      type: string
      pattern: '^([01]\d|2[0-3]):[0-5]\d$'
      examples: ['09:30']
...
    AppointmentStatus:
      type: string
      enum: [pending, confirmed, checked_in, in_progress, completed, cancelled, no_show]
    AppointmentMutationVersion:
      type: integer
      minimum: 0
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [success, message]
      properties:
        success: { type: boolean, const: false }
        message: { type: string }
        code:
          type: string
          description: Includes APPOINTMENT_VERSION_CONFLICT for stale staff appointment mutations.
        details: { type: object, additionalProperties: true }
...
    TranslationText:
      type: object
      additionalProperties: true
    Translations:
      type: object
      additionalProperties: false
      properties:
        hy: { $ref: '#/components/schemas/TranslationText' }
        ru: { $ref: '#/components/schemas/TranslationText' }
        en: { $ref: '#/components/schemas/TranslationText' }
...
    PriceSnapshot:
      type: object
      additionalProperties: false
      required: [priceType, currency]
      properties:
        priceType: { type: string, enum: [fixed, from, range, on_request] }
        priceFrom:
          anyOf:
            - { type: number, minimum: 0 }
            - { type: 'null' }
        priceTo:
          anyOf:
            - { type: number, minimum: 0 }
            - { type: 'null' }
        currency: { type: string, const: AMD }
    DentistPublicSummary:
      type: object
      additionalProperties: false
      required: [id, firstName, lastName]
      properties:
        id: { $ref: '#/components/schemas/ObjectId' }
        firstName: { type: string }
        lastName: { type: string }
        slug: { $ref: '#/components/schemas/Slug' }
        title: { type: string }
        translations: { $ref: '#/components/schemas/Translations' }
    DentistAdminReference:
      oneOf:
        - { $ref: '#/components/schemas/ObjectId' }
        - type: object
          additionalProperties: false
          required: [_id, firstName, lastName, slug]
          properties:
            _id: { $ref: '#/components/schemas/ObjectId' }
            firstName: { type: string }
            lastName: { type: string }
            slug: { $ref: '#/components/schemas/Slug' }
            title: { type: string }
            translations: { $ref: '#/components/schemas/Translations' }
    ServicePublicSummary:
      type: object
      additionalProperties: false
      required: [id, name, durationMinutes]
      properties:
        id: { $ref: '#/components/schemas/ObjectId' }
        name: { type: string }
        slug: { $ref: '#/components/schemas/Slug' }
        translations: { $ref: '#/components/schemas/Translations' }
        durationMinutes: { type: integer, minimum: 1 }
        priceType: { type: string, enum: [fixed, from, range, on_request] }
        priceFrom:
          anyOf:
            - { type: number, minimum: 0 }
            - { type: 'null' }
        priceTo:
          anyOf:
            - { type: number, minimum: 0 }
            - { type: 'null' }
        currency: { type: string, const: AMD }
    ServiceAdminReference:
      oneOf:
        - { $ref: '#/components/schemas/ObjectId' }
        - type: object
          additionalProperties: false
          required: [_id, name, slug]
          properties:
            _id: { $ref: '#/components/schemas/ObjectId' }
            name: { type: string }
            slug: { $ref: '#/components/schemas/Slug' }
            translations: { $ref: '#/components/schemas/Translations' }
    AvailabilitySlot:
      type: object
      additionalProperties: false
      required: [start, end, startAt, endAt]
      properties:
        start: { $ref: '#/components/schemas/LocalTime' }
        end: { $ref: '#/components/schemas/LocalTime' }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
    Availability:
      type: object
      additionalProperties: false
      required: [date, timezone, available, reason, dentist, service, rules, slots]
      properties:
        date: { $ref: '#/components/schemas/LocalDate' }
        timezone: { type: string, description: Configured clinic IANA timezone. }
        available: { type: boolean }
        reason:
          anyOf:
            - type: string
              enum: [SAME_DAY_BOOKING_DISABLED, CLINIC_CLOSED, DENTIST_NOT_WORKING, NO_COMMON_WORKING_TIME, FULLY_BOOKED]
            - { type: 'null' }
        dentist: { $ref: '#/components/schemas/DentistPublicSummary' }
        service: { $ref: '#/components/schemas/ServicePublicSummary' }
        rules:
          type: object
          additionalProperties: false
          required: [slotIntervalMinutes, bufferMinutes, minBookingNoticeMinutes]
          properties:
            slotIntervalMinutes: { type: integer, enum: [10, 15, 20, 30, 60] }
            bufferMinutes: { type: integer, minimum: 0, maximum: 120 }
            minBookingNoticeMinutes: { type: integer, minimum: 0, maximum: 10080 }
        schedule:
          type: object
          additionalProperties: false
          description: Present when the effective clinic and dentist schedules were evaluated.
          required: [clinicSource, dentistSource, clinicRevision, dentistRevision]
          properties:
            clinicSource: { type: string, enum: [weekly, exception] }
            dentistSource: { type: string, enum: [weekly, exception] }
            clinicRevision: { type: integer, minimum: 0 }
            dentistRevision: { type: integer, minimum: 0 }
        slots:
          type: array
          items: { $ref: '#/components/schemas/AvailabilitySlot' }
    AppointmentPublicResult:
      type: object
      additionalProperties: false
      description: >-
        Safe result returned only to the public booking caller. It intentionally
        excludes phone, email, comments, internal notes, consent evidence,
        mutation history, lock keys, and idempotency internals.
      required: [id, confirmationCode, patientName, date, startTime, endTime, status, dentist, service, price]
      properties:
        id: { $ref: '#/components/schemas/ObjectId' }
        confirmationCode: { type: string, pattern: '^DC-[0-9A-F]{16}$' }
        patientName: { type: string, minLength: 2, maxLength: 120 }
        date: { $ref: '#/components/schemas/LocalDate' }
        startTime: { $ref: '#/components/schemas/LocalTime' }
        endTime: { $ref: '#/components/schemas/LocalTime' }
        status: { $ref: '#/components/schemas/AppointmentStatus' }
        dentist: { $ref: '#/components/schemas/DentistPublicSummary' }
        service: { $ref: '#/components/schemas/ServicePublicSummary' }
        price: { $ref: '#/components/schemas/PriceSnapshot' }
    AppointmentSnapshotPoint:
      type: object
      additionalProperties: false
      required: [dentist, service, date, startTime, endTime]
      properties:
        dentist: { $ref: '#/components/schemas/ObjectId' }
        service: { $ref: '#/components/schemas/ObjectId' }
        date: { $ref: '#/components/schemas/LocalDate' }
        startTime: { $ref: '#/components/schemas/LocalTime' }
        endTime: { $ref: '#/components/schemas/LocalTime' }
    RescheduleHistoryEntry:
      type: object
      additionalProperties: false
      required: [actor, reason, from, to, changedAt]
      properties:
        actor:
          anyOf:
            - { $ref: '#/components/schemas/ObjectId' }
            - { type: 'null' }
        reason: { type: string, maxLength: 500 }
        from: { $ref: '#/components/schemas/AppointmentSnapshotPoint' }
        to: { $ref: '#/components/schemas/AppointmentSnapshotPoint' }
        changedAt: { type: string, format: date-time }
    AppointmentAdmin:
      type: object
      additionalProperties: false
      description: >-
        Authorized administrative appointment. This includes patient contact
        data, internal notes, reschedule history, and server-recorded privacy
        evidence; it is never a public collection schema.
      required:
        - _id
        - confirmationCode
        - patientName
        - patientPhone
        - patientEmail
        - dentist
        - service
        - dentistSnapshot
        - serviceSnapshot
        - priceSnapshot
        - date
        - startTime
        - endTime
        - startAt
        - endAt
        - bufferMinutes
        - mutationVersion
        - scheduleRevision
        - notificationLocale
        - rescheduleHistory
        - status
        - source
        - patientComment
        - internalNote
        - privacyConsentAt
        - privacyConsentMethod
        - privacyPolicyVersion
        - createdAt
        - updatedAt
      properties:
        _id: { $ref: '#/components/schemas/ObjectId' }
        confirmationCode: { type: string, pattern: '^DC-[0-9A-F]{16}$' }
        patientName: { type: string, minLength: 2, maxLength: 120 }
        patientPhone: { type: string, maxLength: 20 }
        patientEmail: { type: string, maxLength: 254 }
        dentist: { $ref: '#/components/schemas/DentistAdminReference' }
        service: { $ref: '#/components/schemas/ServiceAdminReference' }
        dentistSnapshot:
          type: object
          additionalProperties: false
          required: [firstName, lastName, title, translations]
          properties:
            firstName: { type: string }
            lastName: { type: string }
            title: { type: string }
            translations: { $ref: '#/components/schemas/Translations' }
        serviceSnapshot:
          type: object
          additionalProperties: false
          required: [name, durationMinutes, translations]
          properties:
            name: { type: string }
            durationMinutes: { type: integer, minimum: 1 }
            translations: { $ref: '#/components/schemas/Translations' }
        priceSnapshot: { $ref: '#/components/schemas/PriceSnapshot' }
        date: { $ref: '#/components/schemas/LocalDate' }
        startTime: { $ref: '#/components/schemas/LocalTime' }
        endTime: { $ref: '#/components/schemas/LocalTime' }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        bufferMinutes: { type: integer, minimum: 0 }
        mutationVersion: { type: integer, minimum: 0 }
        scheduleRevision:
          type: integer
          minimum: 0
          description: Increments only when the committed appointment occurrence changes.
        notificationLocale:
          description: >-
            Persisted locale for future notifications. Null identifies a legacy
            appointment for which no historical preference was recorded; delivery
            then uses the documented Armenian fallback.
          anyOf:
            - { type: string, enum: [hy, ru, en] }
            - { type: 'null' }
        rescheduleHistory:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/RescheduleHistoryEntry' }
        status: { $ref: '#/components/schemas/AppointmentStatus' }
        source: { type: string, enum: [website, phone, admin] }
        patientComment: { type: string, maxLength: 1000 }
        internalNote: { type: string, maxLength: 2000 }
        privacyConsentAt: { type: string, format: date-time }
        privacyConsentMethod: { type: string, enum: [website, phone, in_person] }
        privacyPolicyVersion:
          type: string
          pattern: '^(?:[0-9]{4}-[0-9]{2}(?:\.[0-9]+)?|legacy-unverified)$'
        createdBy:
          anyOf:
            - { $ref: '#/components/schemas/ObjectId' }
            - { type: 'null' }
        cancelledAt:
          anyOf:
            - { type: string, format: date-time }
            - { type: 'null' }
        cancelledBy:
          anyOf:
            - { $ref: '#/components/schemas/ObjectId' }
            - { type: 'null' }
        cancellationReason: { type: string, maxLength: 500 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Pagination:
      type: object
      additionalProperties: false
      required: [page, limit, total, pages]
      properties:
        page: { type: integer, minimum: 1 }
        limit: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
        pages: { type: integer, minimum: 0 }
...
    PatientBookingFields:
      type: object
      required: [patientName, patientPhone, dentistId, serviceId, date, startTime, privacyAccepted]
      properties:
        patientName: { type: string, minLength: 2, maxLength: 120 }
        patientPhone: { type: string, minLength: 8, maxLength: 30 }
        patientEmail:
          anyOf:
            - { type: string, format: email, maxLength: 254 }
            - { type: string, const: '' }
        dentistId: { $ref: '#/components/schemas/ObjectId' }
        serviceId: { $ref: '#/components/schemas/ObjectId' }
        date: { $ref: '#/components/schemas/LocalDate' }
        startTime: { $ref: '#/components/schemas/LocalTime' }
        patientComment: { type: string, maxLength: 1000 }
        locale:
          type: string
          enum: [hy, ru, en]
          default: hy
          description: Language used for appointment notifications and reminders.
        privacyAccepted:
          type: boolean
          const: true
          description: >-
            Acceptance of the server-configured current privacy policy. The server
            records its own timestamp, method, and policy version; clients do not
            submit those evidence fields.
    PublicBookingRequest:
      allOf:
        - { $ref: '#/components/schemas/PatientBookingFields' }
        - type: object
          required: [challengeToken]
          properties:
            challengeToken:
              type: string
              minLength: 10
              maxLength: 4096
              writeOnly: true
              description: >-
                Required in production for the configured bot-challenge provider.
                It may be omitted only when challenge verification is explicitly
                disabled in a non-production environment.
      unevaluatedProperties: false
    AdminBookingRequest:
      allOf:
        - { $ref: '#/components/schemas/PatientBookingFields' }
        - type: object
          required: [consentMethod]
          properties:
            consentMethod: { type: string, enum: [phone, in_person] }
            source: { type: string, enum: [phone, admin], default: phone }
            internalNote: { type: string, maxLength: 2000 }
      unevaluatedProperties: false
    RescheduleRequest:
      type: object
      additionalProperties: false
      required: [expectedMutationVersion, date, startTime]
      properties:
        expectedMutationVersion: { $ref: '#/components/schemas/AppointmentMutationVersion' }
        date: { $ref: '#/components/schemas/LocalDate' }
        startTime: { $ref: '#/components/schemas/LocalTime' }
        dentistId: { $ref: '#/components/schemas/ObjectId' }
        serviceId: { $ref: '#/components/schemas/ObjectId' }
        reason: { type: string, maxLength: 500 }
    AppointmentStatusRequest:
      type: object
      additionalProperties: false
      required: [expectedMutationVersion, status]
      properties:
        expectedMutationVersion: { $ref: '#/components/schemas/AppointmentMutationVersion' }
        status: { type: string, enum: [confirmed, checked_in, in_progress, completed, no_show] }
        internalNote: { type: string, maxLength: 2000 }
    CancelAppointmentRequest:
      type: object
      additionalProperties: false
      required: [expectedMutationVersion, reason]
      properties:
        expectedMutationVersion: { $ref: '#/components/schemas/AppointmentMutationVersion' }
        reason: { type: string, minLength: 2, maxLength: 500 }
...